+++ src/common/interface/index.ts
@@ -1,4 +1,3 @@
 export * from './async-provider';
 export * from './client';
-export * from './module';
 export * from './remote-config';