+++ src/common/interface/remote-config.ts
@@ -1,11 +1,37 @@
 export interface RemoteConfigClient {
-  init: (...args: any[]) => Promise<any>;
-  close: (...args: any[]) => Promise<void>;
-  sync: (...args: any[]) => Promise<any>;
-  subscribe?: (key: string, callback: (value: any) => any) => Promise<void>;
+  /**
+   * Close the remote configuration center.
+   *
+   * @publicApi
+   */
+  close?: (...args: any[]) => Promise<any>;
+  /**
+   * Periodic synchronization of configuration information (polling).
+   *
+   * @publicApi
+   */
+  sync?: () => Promise<any>;
+  /**
+   * Subscribe to configuration information (callbacks or hot updates).
+   *
+   * @publicApi
+   */
+  subscribe?: (callback: (config: any) => any) => Promise<any>;
+  /**
+   * Use customize arguments to initialize the remote configuration center.
+   *
+   * @publicApi
+   */
+  init: (...args: any[]) => Promise<Partial<any>>;
 }
 
 export interface RemoteConfigInfo<T = any> {
+  /**
+   * Key in the remote configuration center.
+   */
   key: string;
+  /**
+   * Indicates the actual configuration value.
+   */
   value?: T;
 }